requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
pydantic>=2.5.0
PyPDF2>=3.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-groq>=0.2.0
langchain-text-splitters>=0.3.0
sentence-transformers>=3.0.0
supabase>=2.4.0,<2.10.0
python-multipart>=0.0.6
streamlit>=1.40.0
requests>=2.31.0
numpy>=1.24.0
torch>=2.0.0